requirements.txt

Django==1.10.4
django-colorfield==0.1.10
django-ckeditor==5.0.3
pillow
Markdown==2.6.7
bleach==1.5.0
